Full Compression Ratio for Huffman
The dictionary was added to the total size of the file, as this needs to be stored when decompressing. This limits the compression algorithm a little but without it the compression would be useless (as you could not uncompress). I am only looking for ratios but the dictionary affects that.