Run Length Encoding
The run length encoding algorithm is added and it returns the compression ratio again as all of my algorithms should be doing. This means I have coded 2 algorithms now (on track). I linked the sources I used for help in the code.

The algorithm was tested on the same 3 test files as the Huffman coding one was tested on.